+++ D5700ComputerEmulator/src/test/kotlin/instructionsTest/ReadTest.kt
@@ -20,7 +20,7 @@ class ReadTest {
         cpu.memory = 0u
         array[0].write(cpu.address.toInt(), 10u)
         read.execute("37", "00")
-        assertEquals(10, cpu.registers[7])
+        assertEquals(10.toUByte(), cpu.registers[7])
     }
 
     @Test
@@ -29,6 +29,6 @@ class ReadTest {
         cpu.memory = 1u
         array[0].write(cpu.address.toInt(), 10u)
         read.execute("37", "00")
-        assertEquals(0, cpu.registers[7])
+        assertEquals(0.toUByte(), cpu.registers[7])
     }
 }
